chatbot_project/requirements.txt
# requirements.txt - Complete Dependencies for Django Chatbot
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
psycopg2-binary==2.9.7
django-redis==5.4.0
requests==2.31.0
python-dotenv==1.0.0

# JWT Authentication
PyJWT==2.8.0
djangorestframework-simplejwt==5.3.0

# Background Tasks
APScheduler==3.10.4
celery==5.3.4
redis==5.0.1

# RAG Pipeline Dependencies
sentence-transformers==2.2.2
faiss-cpu==1.7.4
numpy==1.24.3
torch==2.0.1
transformers==4.33.2

# Utilities
python-decouple==3.8
Pillow==10.0.1
celery
redis
numpy
sentence-transformers
faiss-cpu
